feat: use new tx builder/ network tx builder
