+++ Cargo.toml
@@ -111,13 +111,13 @@ tokio = "1"
 rstest = "0.24.0"
 
 [patch.crates-io]
-# alloy-eips = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-serde = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-signer = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-network = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-provider = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-transport = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-consensus = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-rpc-types-eth = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-rpc-types-engine = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
-# alloy-network-primitives = { git = "https://github.com/alloy-rs/alloy", rev = "2390e6cd5" }
+alloy-eips = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-serde = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-signer = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-network = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-provider = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-transport = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-consensus = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-rpc-types-eth = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-rpc-types-engine = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
+alloy-network-primitives = { git = "https://github.com/lean-apple/alloy.git", branch = "tx-builder-split" }
