+++ crates/rpc-types/src/lib.rs
@@ -18,5 +18,7 @@ pub use receipt::{L1BlockInfo, OpTransactionReceipt, OpTransactionReceiptFields}
 mod transaction;
 pub use transaction::{OpTransactionFields, OpTransactionRequest, Transaction};
 
+mod builder;
+
 pub mod error;
 pub use error::SuperchainDAError;
